+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## Versions
 
 ## 2.1.0
-- Add Guide "The Flame Warden"
+- Add guide "The Flame Warden" for the Alliance
+- Add a short guide for the Horde achievement "The Flame Keeper"
 
 ## 2.0.1
 - Fix Noble Gardener Lua error